infrastructure/requirements.txt
# =============================================================================
# CDK INFRASTRUCTURE DEPENDENCIES
# =============================================================================
aws-cdk-lib~=2.207
constructs>=10.4.2,<11.0.0

# =============================================================================
# DEVELOPMENT DEPENDENCIES (local development & testing)
# =============================================================================
aws-sam-cli~=1.142

black~=25.1
flake8~=7.3
isort~=6.0
mypy~=1.17
pre-commit~=4.2

httpx~=0.28
python-dotenv~=1.1

boto3~=1.39
botocore~=1.39

email-validator~=2.2
uvicorn[standard]~=0.35
